# wms-contract/src/main/resources/openapi/wms-v1.yaml
openapi: 3.1.0
info:
  title: WMS v1 HTTP contract
  version: 0.1.0
  description: |
    公开前缀 `/api/wms/v1`，内部前缀 `/internal/wms/v1`。
    enterpriseId/operator 来自已验证 OIDC token；请求中的企业字段仅核验一致性，不能扩大权限。
    数量 JSON 为十进制字符串。时间 RFC3339 UTC。列表 cursor 分页，默认 50、上限 200。
    写入头必须带 Authorization、Idempotency-Key、X-Request-Id；clientOperationId 若出现必须与 Idempotency-Key 一致。
    Seata TCC Try/Confirm/Cancel 不是 REST 资源，不提供自定义 /prepare 或 /decisions。
    OIDC issuer/client/权限映射由环境变量配置，本文件不含密钥。
    本契约覆盖 S1 主数据与后续切片接口形状；未实现切片不得把 200 当成业务已交付。
servers:
  - url: /
security:
  - oidc: []
...
paths:
...
  /api/wms/v1/warehouses/{warehouseId}/tasks/{taskId}/action-effects:
    get:
      operationId: listTaskActionEffects
      tags: [idempotency]
      summary: 恢复动作槽与effect身份
      security:
        - oidc: [task.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/jobs/{jobId}:
    get:
      operationId: getJob
      tags: [job]
      summary: 作业进度
      security:
        - oidc: [job.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/jobs/{jobId}/retries:
    post:
      operationId: retryJob
      tags: [job]
      summary: 失败分片恢复同一业务任务
      security:
        - oidc: [job.retry]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRetryRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/warehouses/{warehouseId}/action-effects:
    post:
      operationId: registerActionEffect
      tags: [idempotency]
      summary: 按权威事实登记effect身份，换客户端键仍返回同一effect
      security:
        - oidc: [task.read]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionEffectRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/action-effects/{effectId}:
    get:
      operationId: getActionEffect
      tags: [idempotency]
      summary: 查询active/applied命令与safeToRetry
      security:
        - oidc: [task.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/EffectId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/action-effects/{effectId}/execution-attempts:
    post:
      operationId: createExecutionAttempt
      tags: [idempotency]
      summary: 安全重授权；未知状态不得直接重做
      security:
        - oidc: [task.claim]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/EffectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionAttemptRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses:
    post:
      operationId: createWarehouse
      tags: [masterdata]
      summary: 创建仓库
      security:
        - oidc: [masterdata.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreateRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    get:
      operationId: listWarehouses
      tags: [masterdata]
      summary: 可访问仓列表
      security:
        - oidc: [masterdata.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}:
    get:
      operationId: getWarehouse
      tags: [masterdata]
      summary: 仓库详情
      security:
        - oidc: [masterdata.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: issuer 与 client 由 WMS_OIDC_ISSUER / WMS_OIDC_CLIENT_ID 指向已有提供方，仓库不写死密钥
      openIdConnectUrl: https://issuer.example.invalid/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
      description: 与 clientOperationId 一致；同键同内容返回原结果，同键异内容 409
    XRequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    Cursor:
      description: 不透明游标，绑定原查询条件；更换筛选或投影世代后从第一页重新读取。
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    WarehouseId:
      name: warehouseId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    JobId:
      name: jobId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    EffectId:
      name: effectId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
  responses:
...
    Error:
      description: 业务或协议错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 64
...
    Version:
      type: integer
      format: int64
      minimum: 0
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, retryable]
      properties:
        code:
          type: string
          enum:
            - INVALID_QUANTITY
            - INVALID_UNIT
            - REQUIRED_LOT
            - INVALID_EXPIRY
            - UNAUTHENTICATED
            - WAREHOUSE_FORBIDDEN
            - RESOURCE_NOT_FOUND
            - IDEMPOTENCY_PAYLOAD_MISMATCH
            - VERSION_CONFLICT
            - STOCK_INSUFFICIENT
            - STOCK_FROZEN
            - EXPIRED_STOCK
            - SERIAL_OWNERSHIP_CONFLICT
            - STALE_OWNER_EPOCH
            - TCC_TRANSACTION_PENDING
            - INVALID_STATE
            - TCC_CONTEXT_MISMATCH
            - TCC_OWNER_CONFLICT
            - ROUTE_EPOCH_STALE
            - RESOURCE_BUDGET_EXCEEDED
            - DEPENDENCY_UNAVAILABLE
            - EFFECT_IDENTITY_CONFLICT
            - STALE_EXECUTION_ATTEMPT
            - EFFECT_ALREADY_APPLIED
            - RECOVERY_PENDING
            - INVALID_ARGUMENT
            - INVALID_PAGE
            - DATABASE_UNAVAILABLE
            - QUERY_OVERLOADED
            - RATE_LIMITED
            - DUPLICATE_DOCUMENT
            - DUPLICATE_INSPECTION
        message: { type: string }
        requestId: { type: string }
        operationId: { $ref: '#/components/schemas/Id' }
        retryable: { type: boolean }
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field: { type: string }
              code: { type: string }
    CursorPage:
      type: object
      additionalProperties: false
      required: [items, limit]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ResourceEnvelope' }
        nextCursor: { type: string }
        limit: { type: integer, minimum: 1, maximum: 200 }
        asOf: { type: string, format: date-time }
        lagSeconds: { type: integer, minimum: 0 }
    ResourceEnvelope:
      type: object
      additionalProperties: true
      required: [id, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { $ref: '#/components/schemas/Version' }
        status: { type: string }
        actions:
          type: array
          items: { type: string }
          description: 交互提示，写入仍以服务端校验为准
    AcceptedOperation:
      type: object
      additionalProperties: false
      required: [operationId, status, statusUrl]
      properties:
        operationId: { $ref: '#/components/schemas/Id' }
        sourceOperationId: { $ref: '#/components/schemas/Id' }
        commandId: { $ref: '#/components/schemas/Id' }
        fulfillmentId: { $ref: '#/components/schemas/Id' }
        receiptId: { $ref: '#/components/schemas/Id' }
        executionId: { $ref: '#/components/schemas/Id' }
        physicalStatus: { type: string }
        stockSyncStatus: { type: string }
        status: { type: string }
        version: { $ref: '#/components/schemas/Version' }
        statusUrl: { type: string }
        safeToRetry: { type: boolean }
...
    JobRetryRequest:
      type: "object"
      additionalProperties: false
      properties:
        action:
          type: "string"
          maxLength: 64
          pattern: "(?i)RECLAIM|CLAIM|TAKEOVER"
...
    ActionEffectRequest:
      type: object
      additionalProperties: false
      required: [factType, factParentId, factPartId, factLineId, action]
      properties:
        factType: { type: string }
        factParentId: { $ref: '#/components/schemas/Id' }
        factPartId: { $ref: '#/components/schemas/Id' }
        factLineId: { $ref: '#/components/schemas/Id' }
        action: { type: string }
        digestVersion: { $ref: '#/components/schemas/Version' }
        clientOperationId: { $ref: '#/components/schemas/Id' }
    ExecutionAttemptRequest:
      type: "object"
      additionalProperties: false
      required: ["expectedEffectVersion"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        expectedEffectVersion:
          type: "integer"
          minimum: 0
        digestVersion:
          type: "integer"
          minimum: 0
        previousCommandId:
          type: "string"
          maxLength: 64
    WarehouseCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["code","name","timezone"]
      properties:
        clientOperationId:
          type: "string"
          maxLength: 64
        code:
          type: "string"
          maxLength: 32
          minLength: 1
        name:
          type: "string"
          maxLength: 128
          minLength: 1
        timezone:
          type: "string"
          maxLength: 64
          minLength: 1
...